refactor predict and redis controllers to use new wrappers.
merged the two controllers into a single one.
