+++ server/controllers/index.js
@@ -1,11 +1,9 @@
 import modelController from './model.controller';
 import predictController from './predict.controller';
-import redisController from './redis.controller';
 import uploadController from './upload.controller';
 
 export default {
   modelController,
   predictController,
-  redisController,
   uploadController,
 };
