+++ server/controllers/predict.controller.js
@@ -1,11 +1,14 @@
 import httpStatus from 'http-status';
 import uuidv4 from 'uuid/v4';
-import { promisify } from 'util';
-import client from '../config/redis';
+import redis from '../config/redis';
 import config from '../config/config';
 import logger from '../config/winston';
 
 // helper functions
+function isArray(a) {
+  return (!!a) && (a.constructor === Array);
+}
+
 function isValidPredictdata(data) {
   const requiredKeys = [
     // 'modelName',
@@ -20,55 +23,56 @@ function isValidPredictdata(data) {
   return true;
 }
 
-async function addRedisKey(client, redisKey, data) {
-  const hmsetAsync = promisify(client.hmset).bind(client);
-  const now = new Date().toISOString();
-  try {
-    const response = await hmsetAsync([
-      redisKey,
-      'original_name', data.imageName, // to save results with the same name
-      'input_file_name', data.uploadedName || data.imageName, // used for unique files
-      'model_name', data.modelName || '',
-      'model_version', data.modelVersion || '',
-      'postprocess_function', data.postprocessFunction || '',
-      'preprocess_function', data.preprocessFunction || '',
-      'cuts', data.cuts || '0', // to split up very large images
-      'url', data.imageUrl || '', // unused?
-      'scale', data.dataRescale || '',
-      'label', data.dataLabel || '',
-      'status', 'new',
-      'created_at', now,
-      'updated_at', now,
-      'identity_upload', config.hostname,
-    ]);
-    logger.debug(`"HMSET ${redisKey}" response: ${response}`);
-    return response;
-  } catch (err) {
-    logger.error(`Encountered error during "HMSET ${redisKey}": ${err}`);
-    throw err;
-  }
-}
-
-async function lpush(client, queueName, redisKey) {
-  const lpushAsync = promisify(client.lpush).bind(client);
-  try {
-    let response;
-    if (Array.isArray(redisKey)) {
-      response = await lpushAsync(queueName, ...redisKey);
-    } else {
-      response = await lpushAsync(queueName, redisKey);
-    }
-    logger.debug(`"LPUSH ${redisKey}" response: ${response}`);
-    return response;
-  } catch (err) {
-    logger.error(`Encountered error during "LPUSH ${queueName} ${redisKey}": ${err}`);
-    throw err;
-  }
-}
-
 // route handlers
 async function getJobTypes(req, res) {
   return res.status(httpStatus.OK).send({ jobTypes: config.jobTypes });
+}
+
+async function getKey(req, res) {
+  const redisHash = req.body.hash;
+  const redisKey = req.body.key;
+  try {
+    let value;
+    if (isArray(redisKey)) {
+      value = await redis.hmget(redisHash, redisKey);
+    } else {
+      value = await redis.hget(redisHash, redisKey);
+    }
+    return res.status(httpStatus.OK).send({ value });
+
+  } catch (err) {
+    logger.error(`Could not get hash ${redisHash} key ${redisKey} values: ${err}`);
+    return res.status(httpStatus.INTERNAL_SERVER_ERROR).send({ message: err });
+  }
+}
+
+async function getJobStatus(req, res) {
+  const redisHash = req.body.hash;
+  const redisKey = 'status';
+  try {
+    const value = await redis.hget(redisHash, redisKey);
+    return res.status(httpStatus.OK).send({ status: value });
+  } catch (err) {
+    logger.error(`Could not get hash ${redisHash} status: ${err}`);
+    return res.status(httpStatus.INTERNAL_SERVER_ERROR).send({ message: err });
+  }
+}
+
+async function expireHash(req, res) {
+  const redisHash = req.body.hash;
+  const expireTime = req.body.expireIn || 3600;
+  try {
+    const value = await redis.expire(redisHash, expireTime);
+    if (parseInt(value) == 0) {
+      logger.warning(`Hash "${redisHash}" not found`);
+      return res.status(httpStatus.NOT_FOUND).send({ value });
+    }
+    logger.debug(`Expiring hash ${redisHash} in ${expireTime} seconds: ${value}`);
+    return res.status(httpStatus.OK).send({ value });
+  } catch (err) {
+    logger.error(`Error during EXPIRE ${redisHash} ${expireTime}: ${err}`);
+    return res.status(httpStatus.INTERNAL_SERVER_ERROR).send({ message: err });
+  }
 }
 
 async function predict(req, res) {
@@ -91,10 +95,28 @@ async function predict(req, res) {
   }
 
   const redisKey = `${queueName}:${req.body.imageName}:${uuidv4()}`;
+  const data = req.body;
+  const now = new Date().toISOString();
 
   try {
-    await addRedisKey(client, redisKey, req.body);
-    await lpush(client, queueName, redisKey);
+    const response = redis.hmset(redisKey, [
+      'original_name', data.imageName, // to save results with the same name
+      'input_file_name', data.uploadedName || data.imageName, // used for unique files
+      'model_name', data.modelName || '',
+      'model_version', data.modelVersion || '',
+      'postprocess_function', data.postprocessFunction || '',
+      'preprocess_function', data.preprocessFunction || '',
+      'cuts', data.cuts || '0', // to split up very large images
+      'url', data.imageUrl || '', // unused?
+      'scale', data.dataRescale || '',
+      'label', data.dataLabel || '',
+      'status', 'new',
+      'created_at', now,
+      'updated_at', now,
+      'identity_upload', config.hostname,
+    ]);
+    logger.info(`Added key ${redisKey} with response ${response}`);
+    await redis.lpush(queueName, redisKey);
     return res.status(httpStatus.OK).send({ hash: redisKey });
   } catch (err) {
     return res.status(httpStatus.INTERNAL_SERVER_ERROR).send({ message: err });
@@ -103,5 +125,8 @@ async function predict(req, res) {
 
 export default {
   getJobTypes,
+  getKey,
+  expireHash,
+  getJobStatus,
   predict
 };
